Attempted Merge of Mac's work
This is my attempt to merge Mac's work with the stuff me and Lahom did. A lot of stuff was wrong, though, so I may have missed one or two things. (You may want to check flags, Lahom.) !!!REMEMBER TO WORK WITH THE MOST UPDATED GIT TO AVOID OVERWRITING OTHER PEOPLES WORK!!!