Added west indo-Aryan group back
Indo-Aryan would be too large a group, so I am modifying it to be *much* smaller.  This is the first of five groups.
West, Hindustani, Himalayan, South and East. India is a big place.  Obviously, Dravidians will be their own groups, one for those in the south and the other for those dispersed in the north.